packages/fastApi/requirements.txt
fastapi == 0.103.1
uvicorn == 0.23.2
scikit-learn == 1.3.1
pydantic == 2.3.0
pandas == 2.1.1
numpy == 1.26.0
python-dotenv == 1.0.0
mysql-connector-python == 8.1.0
